requirements.txt
streamlit
streamlit-lottie
requests
pandas
plotly
google-generativeai
python-docx
pdfplumber